requirements.txt
anthropic>=0.40.0
playwright>=1.48.0
pdfplumber>=0.11.0
python-docx>=1.1.0
python-dotenv>=1.0.0
cryptography>=43.0.0
aiohttp>=3.10.0
fastapi>=0.115.0
uvicorn>=0.32.0
jinja2>=3.1.4
aiosqlite>=0.20.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.3.0
rich>=13.9.0
typer>=0.13.0
python-multipart>=0.0.12
